% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getADAPSData}
\alias{getADAPSData}
\title{Function to return adaps_data_all df from NWISWeb or previously retrieved RDB files}
\usage{
getADAPSData(siteNo, StartDt, EndDt, precipSite, dataFile = "")
}
\arguments{
\item{siteNo}{NWIS gaging station id}

\item{precipSite}{NWIS precipitation station id}

\item{StartDt}{a date to start data pulls}

\item{EndDt}{a date to end data pulls}

\item{dataFile}{string of data file path and name}
}
\value{
adaps_data_all data frame containing merged ADAPS data for the requested site and date range
}
\description{
This function accepts an NWIS gage site id, an NWIS precip site id, a StartDate, an EndDate and file names as needed
}
\examples{
\dontrun{
siteNo <- "424314090240601"
StartDt <- '2008-05-30'
EndDt <- '2008-06-15'
getADAPSData(siteNo,StartDt,EndDt,siteNo,dataFile)
}
}

