% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hydrographInteractive.R
\name{hydrographInteractive}
\alias{hydrographInteractive}
\title{Function to return hydrographInteractive html to create interactive hydrograph plot through googleVis}
\usage{
hydrographInteractive(adaps_data_all)
}
\arguments{
\item{adaps_data_all}{data frame containing merged ADAPS data for the requested storm event}
}
\value{
hydrographInteractive html object containing code to create googleVis plot
}
\description{
This function accepts a merged data frame of ADAPS data for a site/storm event
}
\examples{
rdbExample<-rdbExample
hydrographInteractive(rdbExample)
}

