\name{intermediateVolTable}
\alias{intermediateVolTable}
\title{Function to csv file of all intermediate volumes used for calculations}
\usage{
intermediateVolTable(siteNo, StartDt, EndDt, tableOut)
}
\arguments{
  \item{siteNo}{USGS station id}

  \item{StartDt}{start of storm event}

  \item{EndDt}{end of storm event}

  \item{tableOut}{list of data frames of event data}
}
\description{
This function accepts a siteNo, StartDt, EndDt and tableOut
list of data for a given storm event
}

