% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getADAPSData.R
\name{getADAPSData}
\alias{getADAPSData}
\title{Function to return adaps_data_all df from NWISWeb or previously retrieved RDB files}
\usage{
getADAPSData(siteNo, StartDt, EndDt, precipSite, dataFile = "", tzCode = "")
}
\arguments{
\item{siteNo}{NWIS gaging station id}

\item{StartDt}{a date to start data pulls}

\item{EndDt}{a date to end data pulls}

\item{precipSite}{NWIS precipitation station id}

\item{dataFile}{string of data file path and name}

\item{tzCode}{a timezone specification for the data}
}
\value{
adaps_data_all data frame containing merged ADAPS data for the requested site and date range
}
\description{
This function accepts an NWIS gage site id, an NWIS precip site id, a StartDate, an EndDate, a timezone and file names as needed
}
\examples{
\dontrun{
siteNo <- "424314090240601"
StartDt <- '2008-05-30'
EndDt <- '2008-06-15'
getADAPSData(siteNo,StartDt,EndDt,siteNo,dataFile,tzCode="America/Chicago")
}
}

