% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getADAPSData.R
\name{getADAPSData}
\alias{getADAPSData}
\title{Function to return adaps_data_all df from NWISWeb or previously retrieved RDB files}
\usage{
getADAPSData(siteNo, StartDt, EndDt, precipSite, dataFile = "", tzCode = "")
}
\arguments{
\item{siteNo}{NWIS gaging station id}

\item{StartDt}{a date to start data pulls}

\item{EndDt}{a date to end data pulls}

\item{precipSite}{NWIS precipitation station id}

\item{dataFile}{string of data file path and name}

\item{tzCode}{a timezone specification for the data}
}
\value{
adaps_data_all data frame containing merged ADAPS data for the requested site and date range
}
\description{
This function accepts an NWIS gage site id, an NWIS precip site id, a StartDate, an EndDate, a timezone and file names as needed
}
\examples{
\dontrun{
siteNo <- "424421077495301"
StartDt <- '2016-02-03'
EndDt <- '2016-02-03'
precipSite <- "424421077495301"
tzCode <- "America/Jamaica"
adaps_data_all <- getADAPSData(siteNo,StartDt,EndDt,precipSite,tzCode=tzCode)
}
}

