% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hydrographPDF.R
\name{hydrographPDF}
\alias{hydrographPDF}
\title{Function to generate and save pdf version of the hydrograph plot}
\usage{
hydrographPDF(adaps_data_all, siteNo, dateInt)
}
\arguments{
\item{adaps_data_all}{data frame containing merged ADAPS data for the requested storm event}

\item{siteNo}{string containing USGS gage number}

\item{dateInt}{number containing number of hours difference between x-axis tick marks}
}
\description{
This function accepts a merged data frame of ADAPS data for a site/storm event
}

