% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labVolumesTable.R
\name{labVolumesTable}
\alias{labVolumesTable}
\title{Function to generate clean text file of sample amounts for lab}
\usage{
labVolumesTable(StormName, StormStart, StormEnd, tableOut, bottlePickup)
}
\arguments{
\item{StormName}{vector of storm name(s)}

\item{StormStart}{vector of storm start dates}

\item{StormEnd}{vector of storm end dates}

\item{tableOut}{list of data frames containing event data}

\item{bottlePickup}{dates bottles were retrieved}
}
\description{
This function accepts a vector of storm names, vector of storm start and end datetimes and list of dataframes containing event data
}

